pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebooklm-automation"
version = "0.1.0"
description = "Batch prompt runner for NotebookLM study workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "notebooklm-py[browser]>=0.3.4,<0.4.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.11,<0.12",
]

[project.scripts]
nlm-auto = "notebooklm_automation.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]